# backend/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: IUBEO Backend API
  version: 0.2.0
  description: |
    IUBEO のゲームバックエンドの HTTP API と、WebSocket のメッセージ型。

    - HTTP: プレイヤーの発行、自動マッチング、セッションの取得、WebSocket への切り替え
    - WebSocket: `GET /api/v1/sessions/{sessionId}/ws` で切り替えた後のメッセージ。
      型は `components/schemas` の `ClientMessage`(クライアント → サーバー)と
      `ServerMessage`(サーバー → クライアント)。すべて JSON で、`type` で種類を見分ける。

    プレイヤーは HttpOnly Cookie(`iubeo_player`)で識別する。
    状態の正はサーバーで、クライアントは一切判定しない。
servers:
  - url: /
security:
  - playerCookie: []
...
paths:
...
  /api/v1/wisp/token:
    get:
      tags: [wisp]
      operationId: getWispToken
      summary: 実サイトへ出る WISP プロキシに繋ぐための、短い期限つきの URL を発行する
      description: |
        WISP は誰でも使えると、オープンプロキシになる。プレイヤーの Cookie がある人だけに、
        期限つきの署名トークンを付けた WebSocket の URL を返す。WISP 側はトークンを検証してから流す。
      responses:
        "200":
          description: 接続先の URL(トークンを含む)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WispToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /api/v1/sessions:
    post:
      tags: [sessions]
      operationId: createSession
      summary: シングルモードのセッションを作る
      description: マルチプレイのセッションは自動マッチングでのみ作られる。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: 作ったセッション
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    playerCookie:
      type: apiKey
      in: cookie
      name: iubeo_player
...
  responses:
    Unauthorized:
      description: Cookie が無い、または無効
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unavailable:
      description: 今は使えない(設定が無いなど)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 今の状態ではできない(既にセッションに参加中など)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    PlayerId:
      type: string
      example: p1
    SessionId:
      type: string
      example: sess-abc123
    Seq:
      description: セッション内の確定状態の更新順序。単調に増える
      type: integer
      format: int64
      minimum: 0
    Vec3:
      description: "[x, y, z]。ワールド座標(Y-up)"
      type: array
      items:
        type: number
        format: double
      minItems: 3
      maxItems: 3
      example: [1.25, 0, -2.5]
    JsonValue:
      description: kind ごとに中身が決まる自由な JSON(null を含む何でも)
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string

    # ---------- 列挙値 ----------
    SessionMode:
      type: string
      enum: [single, multiplayer]
    SessionStatus:
      type: string
      enum: [waiting, playing, intermission, finished]
    PlayerKind:
      type: string
      enum: [human, cpu]
    ConnectionStatus:
      type: string
      enum: [connecting, connected, disconnected]
    LifeStatus:
      type: string
      enum: [alive, eliminated]
    PhaseStatus:
      type: string
      enum: [active, intermission, completed]
    TaskType:
      type: string
      enum: [read_edit, write, web_search, image_generation]
    TaskStatus:
      type: string
      enum: [pending, completed]
    ItemKind:
      type: string
      enum: [file, lighter]
...
    ObjectKind:
      description: directory 以外は暫定の名前
      type: string
      enum: [directory, workspace, canvas, pc, lighter_stand]
    ObjectScope:
      description: personal は各プレイヤーの区画のもの。shared は全員共通(ディレクトリだけ)
      type: string
      enum: [personal, shared]
    ObjectAvailability:
      type: string
      enum: [available, unavailable]
...
    Outcome:
      type: string
      enum: [victory, defeat]

    # ---------- HTTP ----------
...
    WispToken:
      type: object
      required: [url, expiresAt]
      properties:
        url:
          description: WISP の WebSocket の URL。`token` のクエリを含む
          type: string
          example: wss://example.test/wisp/?token=abc.def
        expiresAt:
          description: トークンの期限(接続を始めるのはこの前に)
          type: string
          format: date-time
...
    CreateSessionRequest:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [single]

    # ---------- アイテム・オブジェクト ----------
    Item:
      description: 手に持つ物。手は 1 つなので、プレイヤーが持てるのは 1 個
      type: object
      required: [id, kind, data]
      properties:
        id:
          type: string
          example: file-02
        kind:
          $ref: "#/components/schemas/ItemKind"
        data:
          description: kind が file なら FileItemData、lighter なら null
          allOf:
            - $ref: "#/components/schemas/JsonValue"
...
    GameObject:
      description: ワールドに置かれた、機能を持つ物体
      type: object
      required: [id, kind, scope, position, users, availability, data]
      properties:
        id:
          type: string
          example: directory-1
        kind:
          $ref: "#/components/schemas/ObjectKind"
        scope:
          $ref: "#/components/schemas/ObjectScope"
        owner:
          description: scope が personal のときだけ
          allOf:
            - $ref: "#/components/schemas/PlayerId"
        position:
          $ref: "#/components/schemas/Vec3"
        users:
          description: 今触っているプレイヤー。personal は最大 1 人、shared は複数人
          type: array
          items:
            $ref: "#/components/schemas/PlayerId"
        availability:
          $ref: "#/components/schemas/ObjectAvailability"
        data:
          description: kind が directory なら DirectoryData。他は未定(null)
          allOf:
            - $ref: "#/components/schemas/JsonValue"
...
    Transform:
      type: object
      required: [position, yaw, pitch, seq]
      properties:
        position:
          description: 足元の位置
          allOf:
            - $ref: "#/components/schemas/Vec3"
        yaw:
          description: ラジアン。0 で -Z を向く
          type: number
          format: double
        pitch:
          description: ラジアン。上向きが正
          type: number
          format: double
        seq:
          description: そのプレイヤーの更新ごとに増やす。増えない更新は捨てる
          type: integer
          format: int64
          minimum: 0
    PlayerStatus:
      description: transform を除くプレイヤー情報(player.updated で配る)
      type: object
      required: [playerId, kind, seat, connection, life, heldItem]
      properties:
        playerId:
          $ref: "#/components/schemas/PlayerId"
        kind:
          $ref: "#/components/schemas/PlayerKind"
        seat:
          description: サンドボックスの区画の番号
          type: integer
          minimum: 1
          maximum: 3
        connection:
          $ref: "#/components/schemas/ConnectionStatus"
        life:
          $ref: "#/components/schemas/LifeStatus"
        heldItem:
          allOf:
            - $ref: "#/components/schemas/Item"
          type: object
          nullable: true
    Player:
      allOf:
        - $ref: "#/components/schemas/PlayerStatus"
        - type: object
          required: [transform]
          properties:
            transform:
              $ref: "#/components/schemas/Transform"

    # ---------- ゲーム進行 ----------
    Task:
      type: object
      required: [taskId, type, assigneePlayerId, targetFileId, status, completedAt]
      properties:
        taskId:
          type: string
        type:
          $ref: "#/components/schemas/TaskType"
        assigneePlayerId:
          $ref: "#/components/schemas/PlayerId"
        targetFileId:
          description: read_edit のときだけ。対象の在庫ファイルの id
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        completedAt:
          description: サーバーが完了を受け付けた時刻
          type: string
          format: date-time
          nullable: true
    Phase:
      type: object
      required: [number, status, startedAt, deadlineAt, tasks]
      properties:
        number:
          type: integer
          minimum: 1
        status:
          $ref: "#/components/schemas/PhaseStatus"
        startedAt:
          type: string
          format: date-time
        deadlineAt:
          description: 残り時間はこれと serverTime から算出する
          type: string
          format: date-time
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
    Team:
      description: チーム共有の勝利条件フラグ
      type: object
      required: [bypassPermission, fireStarted]
      properties:
        bypassPermission:
          type: boolean
        fireStarted:
          type: boolean
    Result:
      type: object
      required: [outcome, decidedAt]
      properties:
        outcome:
          $ref: "#/components/schemas/Outcome"
        decidedAt:
          type: string
          format: date-time
    Game:
      type: object
      required: [phase, team, result]
      properties:
        phase:
          description: 開始前は null
          allOf:
            - $ref: "#/components/schemas/Phase"
          type: object
          nullable: true
        team:
          $ref: "#/components/schemas/Team"
        result:
          description: 決着前は null
          allOf:
            - $ref: "#/components/schemas/Result"
          type: object
          nullable: true
    SessionSnapshot:
      type: object
      required: [schemaVersion, sessionId, mode, status, createdAt, serverTime, seq, players, objects, game]
      properties:
        schemaVersion:
          type: integer
          enum: [2]
        sessionId:
          $ref: "#/components/schemas/SessionId"
        mode:
          $ref: "#/components/schemas/SessionMode"
        status:
          $ref: "#/components/schemas/SessionStatus"
        createdAt:
          type: string
          format: date-time
        serverTime:
          description: 送信時のサーバー時刻。クライアントは時計のずれの補正に使う
          type: string
          format: date-time
        seq:
          $ref: "#/components/schemas/Seq"
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        objects:
          type: array
          items:
            $ref: "#/components/schemas/GameObject"
        game:
          $ref: "#/components/schemas/Game"

    # ---------- WebSocket: クライアント → サーバー ----------
...